python/requirements.txt
pyzmq>=25.1.0
numpy>=1.24.0
pyserial>=3.5
